requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==2.3.4
python-telegram-bot==20.8
python-dotenv==1.0.1
lxml==4.9.3